Signup form button function edited
