+++ src/pages/FindStudy/FindMain.tsx
