+++ src/pages/FindStudy/FindStudy.tsx
