+++ src/pages/Users/Login.tsx
@@ -1,35 +1,36 @@
-import { FunctionComponent, useState } from "react";
+import React, { useState } from "react";
 import styled from "styled-components";
+import { Link } from "react-router-dom";
 import { useForm } from "react-hook-form";
 import { FaEyeSlash, FaEye } from "react-icons/fa";
-import PageTitle from "../components/PageTitle";
-import BackBtn from "../components/BackBtn";
+import PageTitle from "../../components/Users/PageTitle";
+import {
+  Button,
+  GoogleButton,
+  KakaoButton,
+} from "../../components/Users/Button";
 
-interface ISigninFormValues {
+interface ILoginForm {
   email: string;
   password: string;
-  confirmPassword: string;
 }
 
-function EmailSignin({ email, password, confirmPassword }: ISigninFormValues) {
+function Login({ email, password }: ILoginForm) {
   const {
     register,
     handleSubmit,
     formState: { errors },
-    setError,
-  } = useForm<ISigninFormValues>();
+  } = useForm<ILoginForm>();
 
-  const onSubmit = (data: ISigninFormValues) => {
-    if (data.password !== data.confirmPassword) {
-      setError("confirmPassword", { message: "비밀번호가 일치하지 않습니다." });
-    }
+  const onSubmit = (data: ILoginForm) => {
+    console.log(data);
   };
 
+  // Password visibility toggle
   const [passwordType, setPasswordType] = useState({
     type: "password",
     visible: false,
   });
-
   const handleVisibility = (event: any) => {
     setPasswordType(() => {
       if (!passwordType.visible) {
@@ -38,18 +39,16 @@ function EmailSignin({ email, password, confirmPassword }: ISigninFormValues) {
       return { type: "password", visible: false };
     });
   };
-
   return (
     <Container>
       <Wrapper>
-        <BackBtn to="/signin" />
         <PageTitle
-          title="회원가입"
-          expFirst="반갑습니다."
-          expSecond="열정적인 멤버들이 기다리고 있어요!"
+          title="로그인"
+          expFirst="환영합니다."
+          expSecond="로그인 후 스터딧해보세요!"
         />
-        <SigninForm onSubmit={handleSubmit(onSubmit)}>
-          <div className="input-wrapper email-input">
+        <LoginForm onSubmit={handleSubmit(onSubmit)}>
+          <div className="input-wrapper">
             <label htmlFor="email">이메일</label>
             <input
               placeholder="example@studyit.com"
@@ -65,7 +64,8 @@ function EmailSignin({ email, password, confirmPassword }: ISigninFormValues) {
                   ? { border: "1px solid #f36355" }
                   : { border: "1px solid #e4e7ec" }
               }
-            />
+              name="email"
+            ></input>
             <p className="error-message">{errors?.email?.message}</p>
           </div>
           <div className="input-wrapper">
@@ -92,34 +92,23 @@ function EmailSignin({ email, password, confirmPassword }: ISigninFormValues) {
               {passwordType.visible ? <EyeOpen /> : <EyeClose />}
             </span>
           </div>
-          <p className="help-text">영문/숫자/특수문자 조합, 8자~32자</p>
           <p className="error-message">{errors?.password?.message}</p>
-          <div className="input-wrapper">
-            <label htmlFor="confirmPassword">비밀번호 확인</label>
-            <input
-              type={passwordType.type}
-              placeholder="비밀번호를 한번 더 입력해주세요"
-              {...register("confirmPassword")}
-              style={
-                errors.confirmPassword
-                  ? { border: "1px solid #f36355" }
-                  : { border: "1px solid #e4e7ec" }
-              }
-              name="confirmPassword"
-            />
-            <span onClick={handleVisibility}>
-              {passwordType.visible ? <EyeOpen /> : <EyeClose />}
-            </span>
+          <div className="button-wrapper">
+            <Button text="이메일로 로그인하기" />
+            <KakaoButton text="카카오로 로그인하기" />
+            <GoogleButton text="구글로 로그인하기" />
           </div>
-          <p className="error-message">{errors?.confirmPassword?.message}</p>
-          <button type="submit">회원가입하기</button>
-        </SigninForm>
+        </LoginForm>
+        <UtilityBox>
+          <StyledLink to="/signin">아직 회원이 아니신가요?</StyledLink>
+          <StyledLink to="/find-password">비밀번호 찾기</StyledLink>
+        </UtilityBox>
       </Wrapper>
     </Container>
   );
 }
 
-export default EmailSignin;
+export default Login;
 
 const Container = styled.div`
   width: 100%;
@@ -137,16 +126,9 @@ const Wrapper = styled.div`
   justify-content: center;
 `;
 
-const SigninForm = styled.form`
+const LoginForm = styled.form`
   display: flex;
   flex-direction: column;
-
-  .help-text {
-    display: inline-block;
-    color: ${(props) => props.theme.grayColors.gray500};
-    font-size: 14px;
-    margin-top: 6px;
-  }
 
   .input-wrapper {
     position: relative;
@@ -162,17 +144,17 @@ const SigninForm = styled.form`
     input {
       width: 400px;
       height: 48px;
-      border: 1px solid #e0edef;
+      border: 1px solid ${(props) => props.theme.grayColors.gray200};
       border-radius: 6px;
       padding-left: 15px;
       font-size: 16px;
       color: ${(props) => props.theme.grayColors.gray900};
 
-      :focus {
+      &:focus {
         outline: none;
       }
 
-      ::placeholder {
+      &::placeholder {
         color: #98a2b3;
         font-size: 16px;
       }
@@ -181,27 +163,15 @@ const SigninForm = styled.form`
     &:first-of-type {
       margin-bottom: 20px;
     }
-
-    &:nth-of-type(3) {
-      margin-top: 20px;
-    }
   }
-  button {
-    width: 400px;
-    height: 48px;
-    background-color: ${(props) => props.theme.primaryColor};
-    opacity: 50%;
-    border-radius: 6px;
-    border: none;
-    color: #fff;
-    font-size: 16px;
+
+  .button-wrapper {
     margin-top: 40px;
-    cursor: pointer;
   }
 
   .error-message {
-    font-size: 14px;
     margin-top: 6px;
+    font-size: 14px;
     color: ${(props) => props.theme.alertColors.error.text};
   }
 `;
@@ -222,4 +192,16 @@ const EyeClose = styled(FaEyeSlash)`
   position: absolute;
   bottom: 14px;
   right: 12px;
+`;
+
+const UtilityBox = styled.div`
+  width: 400px;
+  display: flex;
+  justify-content: space-between;
+  margin-top: 15px;
+`;
+
+const StyledLink = styled(Link)`
+  font-size: 14px;
+  color: ${(props) => props.theme.grayColors.gray500};
 `;
